+++ src/app/middlewares/setPermissions.ts
@@ -1,44 +1,54 @@
 /**
- * プロジェクトメンバー権限セットミドルウェア
+ * 権限セットミドルウェア
  */
 import * as cinerino from '@cinerino/domain';
-import * as createDebug from 'debug';
 import { NextFunction, Request, Response } from 'express';
 import * as mongoose from 'mongoose';
-
-import { IRole, RoleName } from '../iam';
-
-// tslint:disable-next-line:no-require-imports no-var-requires
-const roles: IRole[] = require('../../../roles.json');
-
-const debug = createDebug('cinerino-api:middlewares');
 
 const RESOURCE_SERVER_IDENTIFIER = <string>process.env.RESOURCE_SERVER_IDENTIFIER;
 
 export default async (req: Request, _: Response, next: NextFunction) => {
     try {
+        let isProjectMember = false;
         let memberPermissions: string[] = [];
-        const customerPermissions: string[] = [];
+
+        const memberRepo = new cinerino.repository.Member(mongoose.connection);
+        const roleRepo = new cinerino.repository.Role(mongoose.connection);
 
         // プロジェクトが決定していれば権限をセット
         if (req.project !== undefined && req.project !== null && typeof req.project.id === 'string') {
-            memberPermissions = await fixMemberPermissions(req)({
-                member: new cinerino.repository.Member(mongoose.connection)
+            // プロジェクト決定済のリクエストに対してプロジェクトメンバー権限を決定する
+            memberPermissions = await cinerino.service.iam.searchPermissions({
+                project: { id: req.project.id },
+                member: { id: req.user.sub }
+            })({
+                member: memberRepo,
+                role: roleRepo
             });
-            debug('project member permissions fixed.', memberPermissions);
+            memberPermissions = memberPermissions.map((p) => `${RESOURCE_SERVER_IDENTIFIER}/${p}`);
 
-            // プロジェクトメンバーでない場合、`customer`ロールに設定
             if (memberPermissions.length === 0) {
-                const role = roles.find((r) => r.roleName === RoleName.Customer);
-                if (role !== undefined) {
-                    customerPermissions.push(...role.permissions.map((p) => `${RESOURCE_SERVER_IDENTIFIER}/${p}`));
-                }
+                // プロジェクトメンバーが見つからない場合、アプリケーションクライアントとして権限検索
+                memberPermissions = await cinerino.service.iam.searchPermissions({
+                    project: { id: req.project.id },
+                    member: { id: req.user.client_id }
+                })({
+                    member: memberRepo,
+                    role: roleRepo
+                });
+                memberPermissions = memberPermissions.map((p) => `${RESOURCE_SERVER_IDENTIFIER}/${p}`);
             }
+
+            isProjectMember = await checkProjectMember({
+                project: { id: req.project.id },
+                member: { id: req.user.sub }
+            })({
+                member: memberRepo
+            });
         }
 
-        req.customerPermissions = customerPermissions;
         req.memberPermissions = memberPermissions;
-        req.isProjectMember = Array.isArray(req.memberPermissions) && req.memberPermissions.length > 0;
+        req.isProjectMember = isProjectMember;
 
         next();
     } catch (error) {
@@ -46,33 +56,28 @@ export default async (req: Request, _: Response, next: NextFunction) => {
     }
 };
 
-/**
- * プロジェクト決定済のリクエストに対してプロジェクトメンバー権限を決定する
- */
-function fixMemberPermissions(req: Request) {
+function checkProjectMember(params: {
+    project: { id: string };
+    member: { id: string };
+}) {
     return async (repos: {
         member: cinerino.repository.Member;
-    }): Promise<string[]> => {
-        let permissions: string[] = [];
+    }) => {
+        let isMember = false;
 
-        if (req.project !== undefined && req.project !== null && typeof req.project.id === 'string') {
-            const projectMembers = await repos.member.search({
-                project: { id: { $eq: req.project.id } },
-                member: { id: { $eq: req.user.sub } }
-            });
+        const members = await repos.member.search({
+            project: { id: { $eq: params.project.id } },
+            member: { id: { $eq: params.member.id } }
+        });
+        if (members.length > 0) {
+            const member = members[0];
 
-            projectMembers.forEach((projectMember) => {
-                projectMember.member.hasRole.forEach((memberRole) => {
-                    const role = roles.find((r) => r.roleName === memberRole.roleName);
-                    if (role !== undefined) {
-                        permissions.push(...role.permissions.map((p) => `${RESOURCE_SERVER_IDENTIFIER}/${p}`));
-                    }
-                });
-            });
-
-            permissions = [...new Set(permissions)];
+            // メンバータイプが`Person`かつロールを持っていればプロジェクトメンバー
+            isMember = member.member.typeOf === cinerino.factory.personType.Person
+                && Array.isArray(member.member.hasRole)
+                && member.member.hasRole.length > 0;
         }
 
-        return permissions;
+        return isMember;
     };
 }